adjusts the offsets of multi staged (and summarized) code examples
