+++ test/edu/ucsc/refactor/internal/changers/ChangersTest.java
@@ -1,6 +1,5 @@
 package edu.ucsc.refactor.internal.changers;
 
-import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Iterables;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
@@ -29,7 +28,6 @@ import java.net.URL;
 import java.net.URLEncoder;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
 
 import static org.hamcrest.CoreMatchers.is;
@@ -1692,67 +1690,7 @@ public class ChangersTest {
     }
 
 
-    @Test public void testClipSpaceGeneration() throws Exception {
-        final Source src = InternalUtil.createQuickSortSource();
 
-        final List<Clip> clipSpace = makeClipSpace(src, Vesper.createIntrospector());
-
-        assertThat(clipSpace.isEmpty(), is(false));
-
-    }
-
-    @Test public void testDeleteDifferences() throws Exception {
-        final Source src = InternalUtil.createQuickSortSource();
-
-        final Introspector introspector = Vesper.createIntrospector();
-        final List<Clip> clipSpace = makeClipSpace(src, introspector);
-        final List<Clip> clipOneToBaseClip  = ImmutableList.of(clipSpace.get(clipSpace.size() -
-                1), clipSpace.get(1));
-
-        final Diff diff = introspector.differences(
-                clipOneToBaseClip.get(0).getSource(),
-                clipOneToBaseClip.get(1).getSource()
-        );
-
-        assertThat(diff.getChangesFromOriginal().isEmpty(), is(true));
-        assertThat(diff.getDeletesFromOriginal().isEmpty(), is(false));
-    }
-
-
-    @Test public void testInsertDifferences() throws Exception {
-        final Source src = InternalUtil.createQuickSortSource();
-
-        final Introspector introspector = Vesper.createIntrospector();
-        final List<Clip> clipSpace = makeClipSpace(src, introspector);
-
-        final Diff diff = introspector.differences(
-                clipSpace.get(1).getSource(),
-                clipSpace.get(2).getSource()
-        );
-
-        assertThat(diff.getChangesFromOriginal().isEmpty(), is(true));
-        assertThat(diff.getInsertsFromOriginal().isEmpty(), is(false));
-    }
-
-    @Test public void testChangeDifferences() throws Exception {
-        final Source src = InternalUtil.createQuickSortSource();
-
-        final Introspector introspector = Vesper.createIntrospector();
-        final List<Clip> clipSpace = makeClipSpace(src, introspector);
-
-        final Source a = clipSpace.get(0).getSource();
-        final Source b = clipSpace.get(1).getSource();
-
-        final Source c = Source.from(b, b.getContents().replaceAll("swap", "exchange"));
-
-        final Diff diff = introspector.differences(
-                a,
-                c
-        );
-
-        assertThat(diff.getChangesFromOriginal().isEmpty(), is(false));
-        assertThat(diff.getInsertsFromOriginal().isEmpty(), is(true));
-    }
 
 
     @Test public void testWrappingOfIncompleteExample() throws Exception {
@@ -2004,111 +1942,7 @@ public class ChangersTest {
         in.close();
     }
 
-    @Test public void testClipSpaceForwardPatching() throws Exception {
-        final Source src = InternalUtil.createQuickSortSource();
 
-        final Introspector introspector = Vesper.createIntrospector();
-        final List<Clip> clipSpace = makeClipSpace(src, introspector);
-
-        final Clip clip = Clip.sync(
-                introspector,
-                clipSpace.subList(1, clipSpace.size())
-        );
-
-        assertThat(clip != null, is(true));
-
-        assert clip != null;
-        final Source patched = clip.getSource();
-
-        final String expected  = clipSpace.get(clipSpace.size() - 1).getSource().getContents();
-        final String revised   = patched.getContents();
-
-
-        assertThat(revised.equals(expected), is(true));
-    }
-
-    @Test public void testSyncingClipAfterBaseWithBaseClip() throws Exception {
-        final Source src = InternalUtil.createQuickSortSource();
-
-        final Introspector introspector = Vesper.createIntrospector();
-        final List<Clip> clipSpace = makeClipSpace(src, introspector);
-        final List<Clip> clipOneToBaseClip  = clipSpace.subList(0, 2);
-
-        final Clip   clip      = Clip.sync(introspector, clipOneToBaseClip);
-
-        assert clip != null;
-        final Source patched   = clip.getSource();
-        final String expected  = clipOneToBaseClip.get(clipOneToBaseClip.size() - 1).getSource().getContents();
-        final String revised   = patched.getContents();
-        assertThat(revised.equals(expected), is(true));
-    }
-
-
-    @Test public void testSummarizeSingleSourceCode() throws Exception {
-        final Source src = InternalUtil.createQuickSortSource();
-
-        final Introspector introspector = Vesper.createIntrospector();
-        List<Location> foldingLocations = introspector.summarize("quicksort", src, 17);
-        assertThat(foldingLocations.isEmpty(), is(false));
-    }
-
-    @Test public void testSummarizeSingleMethodAndLongSourceCode() throws Exception {
-        final Source src = InternalUtil.createSourceWithShortNameMembers();
-
-        final Introspector introspector = Vesper.createIntrospector();
-        List<Location> foldingLocations = introspector.summarize("qsort", src, 17);
-        assertThat(foldingLocations.isEmpty(), is(false));
-
-    }
-
-
-    @Test public void testSummarizeSourceCodeWithStaticNestedClass() throws Exception {
-        final Source src = InternalUtil.createSourceWithStaticNestedClass_ClippingEntireInnerClass();
-
-        final Introspector introspector = Vesper.createIntrospector();
-        List<Location> foldingLocations = introspector.summarize("main", src, 17);
-        assertThat(foldingLocations.isEmpty(), is(false));
-
-    }
-
-    @Test public void testSummarizeAllPossibleClips() throws Exception {
-        final Source src = InternalUtil.createQuickSortSource();
-
-        final Introspector introspector = Vesper.createIntrospector();
-        Map<Clip, List<Location>> allSummaries = introspector.summarize(introspector.multiStage
-                (src), 17);
-        assertThat(allSummaries.isEmpty(), is(false));
-    }
-
-
-    @Test public void testSummarizeClipByRanking() throws Exception {
-        final Source src = InternalUtil.createQuickSortSource();
-
-        final Introspector introspector = Vesper.createIntrospector();
-        List<Clip> clips = introspector.multiStage(src);
-
-        final Clip clip = Clip.find(src, clips);
-        assertNotNull(clip);
-
-        List<Location> foldingLocations = introspector.summarize(clip, 17);
-        assertThat(foldingLocations.isEmpty(), is(false));
-    }
-
-
-
-    @Test public void testMultiStageBrokenCode() throws Exception {
-        final Source src = InternalUtil.createGeneralBrokenSource();
-
-        final Introspector introspector = Vesper.createIntrospector();
-        List<Clip> clips = introspector.multiStage(src);
-
-        assertThat(clips.isEmpty(), is(false));
-    }
-
-
-    private static List<Clip> makeClipSpace(Source src, Introspector introspector){
-        return introspector.multiStage(src);
-    }
 
     private static void  checkChangeCreation(SourceChanger changer, Edit resolved){
        checkChangeCreation(changer, resolved, true);
